Improved codebeat score for builders :recycle:
