+++ Source/Tag/ID32TagCommonStringFramesBuilder.swift
@@ -80,30 +80,4 @@ public class ID32TagCommonStringFramesBuilder {
         frames[.subtitle] = frame
         return self
     }
-
-    public func genre(frame: ID3FrameGenre) -> Self {
-        frames[.genre] = frame
-        return self
-    }
-
-    public func discPosition(frame: ID3FramePartOfTotal) -> Self {
-        frames[.discPosition] = frame
-        return self
-    }
-
-    public func trackPosition(frame: ID3FramePartOfTotal) -> Self {
-        frames[.trackPosition] = frame
-        return self
-    }
-
-    public func attachedPicture(pictureType: ID3PictureType, frame: ID3FrameAttachedPicture) -> Self {
-        frames[.attachedPicture(pictureType)] = frame
-        return self
-    }
-
-    public func unsynchronisedLyrics(language: ID3FrameContentLanguage,
-                                     frame: ID3FrameUnsynchronisedLyrics) -> Self {
-        frames[.unsynchronizedLyrics(language)] = frame
-        return self
-    }
 }
